LayoutTests/fast/mediastream/constructors.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("Tests the JSEP PeerConnection related constructors.");

shouldBeTrue("typeof webkitPeerConnection00 === 'function'");
shouldBeTrue("typeof SessionDescription === 'function'");
shouldBeTrue("typeof IceCandidate === 'function'");
shouldBeTrue("typeof webkitRTCPeerConnection === 'function'");
shouldBeTrue("typeof RTCSessionDescription === 'function'");
shouldBeTrue("typeof RTCIceCandidate === 'function'");

shouldThrow("webkitPeerConnection00()");
shouldThrow("SessionDescription()");
shouldThrow("IceCandidate()");
shouldThrow("webkitRTCPeerConnection()");
shouldThrow("RTCSessionDescription()");
shouldThrow("RTCIceCandidate()");

shouldNotThrow("new webkitPeerConnection00('STUN foobar.com:12345', function(){});");
shouldNotThrow("new SessionDescription('');");
shouldNotThrow("new IceCandidate('', '');");
shouldNotThrow("new webkitRTCPeerConnection({iceServers:[{url:'stun://foobar.com:12345'}]}, null);");
shouldNotThrow("new RTCSessionDescription({type:'offer',sdp:'foobar'});");
shouldNotThrow("new RTCIceCandidate({candidate:'foo'});");

window.jsTestIsAsync = false;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
